Create semat-LTK-DBC.md
Đoàn Bá Cường Bi9062 - Lê Thúc Kiên Bi9131
The relationship between customer, solution and endeavour can be illustrated like this. The stakeholders identify opportunity while software systems help to address it further. The opportunity focuses on requirements, which demanded by stakeholders. Software systems need to fulfil the requirements for customers to use and consume them. The same opportunity sets up the work. The requirements also guiding the work with scopes and constraints and depend on the requirements, the work makes updates and changes to the software systems to satisfy customers. The team, which is supported by customers, is crucial to perform and plan the work efficiently while applies creative ways of working to yield the best results.